requirements.txt
asgiref==3.6.0
asttokens==2.4.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.6
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
docutils==0.20.1
et-xmlfile==1.1.0
executing==2.0.0
filelock==3.12.0
frozendict==2.3.8
idna==3.4
ipython==8.16.1
jedi==0.19.1
Kivy==2.2.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kivymd==1.1.1
matplotlib-inline==0.1.6
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.39
pure-eval==0.2.2
pyEX==0.5.0
Pygments==2.15.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
requests==2.30.0
six==1.16.0
socketIO-client-nexus==0.7.6
sqlparse==0.4.4
sseclient==0.0.27
stack-data==0.6.3
temporal-cache==0.1.4
traitlets==5.11.2
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.2
virtualenv==20.23.0
wcwidth==0.2.8
websocket-client==1.6.4
